+++ locale/zh-cn/script.lua
@@ -136,6 +136,8 @@ DIAG_REDUNDANT_RETURN_VALUE           =
 '最多只有 {max} 个返回值，但此处返回了第 {rmax} 个值。'
 DIAG_REDUNDANT_RETURN_VALUE_RANGE     =
 '最多只有 {max} 个返回值，但此处返回了第 {rmin} 到第 {rmax} 个值。'
+DIAG_MISSING_RETURN                   =
+'此处需要返回值。'
 
 MWS_NOT_SUPPORT         =
 '{} 目前还不支持多工作目录，我可能需要重启才能支持新的工作目录...'